api/requirements.txt
tensorflow==2.11.0
tensorflow-serving-api==2.11.0
fastapi==0.116.2
uvicorn[standard]
python-multipart
pillow
matplotlib
numpy==1.23.5
